♻️ Refactor use-tts-voice
